--- src/main_menu_scene.hpp
@@ -1,19 +1,0 @@
-#pragma once
-
-#include "event_listener.hpp"
-#include "scene.hpp"
-#include "ui/label.hpp"
-
-struct MainMenuScene : public Scene {
-private:
-    ui::Label m_heading;
-    bool m_should_end{ false };
-    bool m_should_exit{ false };
-
-public:
-    explicit MainMenuScene(ServiceProvider* service_provider);
-
-    [[nodiscard]] UpdateResult update() override;
-    void render(const ServiceProvider& service_provider) override;
-    [[nodiscard]] bool handle_event(const SDL_Event& event) override;
-};
