+++ src/meson.build
@@ -16,16 +16,16 @@ src_files += files(
     'font.hpp',
     'grid.cpp',
     'grid.hpp',
+    'input_event.hpp',
     'input.cpp',
     'input.hpp',
-    'input_event.hpp',
     'key_codes.hpp',
     'magic_enum_wrapper.hpp',
     'main.cpp',
-    'mino.cpp',
-    'mino.hpp',
     'mino_stack.cpp',
     'mino_stack.hpp',
+    'mino.cpp',
+    'mino.hpp',
     'music_manager.cpp',
     'point.hpp',
     'random.cpp',
@@ -38,13 +38,13 @@ src_files += files(
     'sdl_context.hpp',
     'settings.hpp',
     'static_string.hpp',
-    'tetrion.cpp',
-    'tetrion.hpp',
     'tetrion_snapshot.cpp',
     'tetrion_snapshot.hpp',
-    'tetromino.hpp',
+    'tetrion.cpp',
+    'tetrion.hpp',
     'tetromino_type.cpp',
     'tetromino_type.hpp',
+    'tetromino.hpp',
     'text.cpp',
     'text.hpp',
     'types.hpp',
@@ -55,12 +55,12 @@ src_files += files(
     'ui/widget.hpp',
     'ui/layout.hpp',
     'ui/label.hpp',
-    'scene.hpp',
-    'ingame_scene.cpp',
-    'ingame_scene.hpp',
-    'scene.cpp',
-    'main_menu_scene.cpp',
-    'main_menu_scene.hpp',
+    'scenes/scene.hpp',
+    'scenes/ingame/ingame.cpp',
+    'scenes/ingame/ingame.hpp',
+    'scenes/scene.cpp',
+    'scenes/pause/pause.cpp',
+    'scenes/pause/pause.hpp',
     'resource_manager.hpp',
     'scene_id.hpp',
     'service_provider.hpp',
